+++ src/app/products/components/product-list/product-list.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 
 import { ProductService } from './../../services/product.service';
 import { CartService } from './../../../cart/services/cart.service';
-import { IProduct } from '../../models/product.model';
+import { IProduct } from '@models/product.model';
 import { CartPopupService } from 'src/app/cart/services/cart-popup.service';
 
 @Component({