+++ src/app/shared/index.ts
@@ -1,0 +1,6 @@
+export * from './shared.module';
+export * from './material.module';
+export * from './order-by.pipe';
+export * from './click-n-mark.directive';
+export * from './hovered.directive';
+export * from './order-by.pipe';