+++ changelog.md
@@ -66,3 +66,9 @@
 
 8. - login.service uses localStorage to hold its state
    - productsData.json - store products in the store. File is compiled into bundle. Will be redone with http
+
+
+## Task 6
+1. - added json-server
+2. - created api.service to hold urls
+3. - product.service updated to handle api requests